#1002: Work to split up and simplify jrunscript/tools/install module by beginning to remove outdated constructs to separate script
